refactor: Extract UserService interface (ISP + DIP)
- Create UserService interface in service/port package
- Rename UserService to UserServiceImpl in service/adapter
- Implement interface in UserServiceImpl
- Update imports in UserController
