+++ security-module/src/main/java/gon/cue/security/service/adapter/UserServiceImpl.java
@@ -1,7 +1,8 @@
-package gon.cue.security.service;
+package gon.cue.security.service.adapter;
 
 import gon.cue.security.model.User;
 import gon.cue.security.repository.UserRepository;
+import gon.cue.security.service.port.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
@@ -11,35 +12,40 @@ import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
-public class UserService {
+public class UserServiceImpl implements UserService {
 
     private final UserRepository userRepository;
 
+    @Override
     public Optional<User> findById(Long id) {
         return userRepository.findById(id);
     }
 
+    @Override
     public Optional<User> findByUsername(String username) {
         return userRepository.findByUsername(username);
     }
 
+    @Override
     public List<User> findAll() {
         return userRepository.findAll();
     }
 
+    @Override
     public User save(User user) {
         return userRepository.save(user);
     }
 
+    @Override
     public void deleteById(Long id) {
         userRepository.deleteById(id);
     }
 
-    public User updateUser(Long id, User updatedUser) {
+    @Override
+    public User updateUsername(Long id, String newUsername) {
         User user = userRepository.findById(id)
                 .orElseThrow(() -> new UsernameNotFoundException("User not found with id: " + id));
-        user.setUsername(updatedUser.getUsername());
-        user.setPassword(updatedUser.getPassword()); // Consider hashing password if it's updated here
+        user.setUsername(newUsername);
         return userRepository.save(user);
     }
 }
